Add json report field to show when the dns lookup was skipped
This field should clarify when a '.well-known' information with a port
declaration was found, that the dns lookup was skipped.
The field can befound inside the 'DNSResult' object with the name
'Skipped' and is from type boolean.

To give some context:
This change was made based on a Pull-Request inside the 'fed-tester-ui'
project [1] to show a proper message to the user inside the UI
explaining the missing SRV records information.

Relates to issue #104.

Refs:
  - [1] https://github.com/matrix-org/fed-tester-ui/pull/26