Add `gomatrixserverlib.WithWellKnownSRVLookups` to `NewClient`
The API shaped for GMSL in matrix-org/gomatrixserverlib#304 to allow us to control
whether `gomatrixserverlib.Client` performs well-known/SRV lookups or not. So it is
right after all to pass in the vanity name to `GetVersion`, but if the client isn't
configured to look up well-known, it'll do the wrong thing.

Fixes #128.